refactor(structure): make queue header the aggregate entry
