+++ src/core/rw/write_port.hpp
@@ -5,7 +5,7 @@
 #include <cstdint>
 #include <string_view>
 
-#include "spmc_queue.hpp"
+#include "queue.hpp"
 #include "operation.hpp"
 
 namespace LibXR
