+++ src/libxr.hpp
@@ -24,9 +24,6 @@
 #include "lockfree_pool.hpp"
 #include "logger.hpp"
 #include "message.hpp"
-#include "mpmc_queue.hpp"
-#include "spmc_queue.hpp"
-#include "spsc_queue.hpp"
 #if defined(LIBXR_SYSTEM_POSIX_HOST)
 #include "linux_shared_topic.hpp"
 #endif
